service done

+++ install-service.sh
@@ -48,8 +48,10 @@ create_user() {
 install_files() {
     print_status "${BLUE}" "Installing files to ${INSTALL_DIR}"
 
-		ln -s "$SCRIPT_DIR" "$INSTALL_DIR"
-		mkdir -p "$INSTALL_DIR/.n8n"
+		if [ ! -e "$INSTALL_DIR" ]; then
+				ln -s "$SCRIPT_DIR" "$INSTALL_DIR"
+				mkdir -p "$INSTALL_DIR/.n8n"
+		fi
 
     # Set ownership and permissions
     chown -R "${SERVICE_USER}:${SERVICE_GROUP}" "${INSTALL_DIR}"
